requirements.txt
# --- Core ---
jieba
pandas
pyyaml
requests
python-dotenv
tqdm
playwright

# --- AI ---
google-generativeai

# --- NLP (High Precision) ---
ckip-transformers
torch
transformers

# --- Visualization ---
matplotlib
seaborn
wordcloud

# --- Notifications ---
line-bot-sdk>=3.0.0
